Improve module naming and responsibilities
