+++ src/robot/writer/datafilewriter.py
@@ -16,7 +16,8 @@
 from robot.errors import DataError
 from robot.utils import file_writer
 
-from .filewriters import write_to_file
+from .transformers import (SeparatorRemover, SettingCleaner, ForLoopCleaner,
+                           Aligner, Formatter)
 
 
 class DataFileWriter(object):
@@ -39,7 +40,11 @@ class DataFileWriter(object):
         """
 
         with WritingContext(model, **self._options) as ctx:
-            write_to_file(ctx, model)
+            SeparatorRemover().visit(model)
+            SettingCleaner().visit(model)
+            ForLoopCleaner().visit(model)
+            Aligner(ctx).visit(model)
+            Formatter(ctx, RowWriter(ctx)).visit(model)
 
 
 class WritingContext(object):
@@ -103,3 +108,30 @@ class WritingContext(object):
         if format and format not in self._formats:
             raise DataError('Invalid format: %s' % format)
         return format
+
+
+class RowWriter(object):
+
+    def __init__(self, ctx):
+        self.output = ctx.output
+        self.pipes = ctx.pipe_separated
+        self.separator, self.indent_string = self.get_separator_and_indent(ctx)
+
+    def get_separator_and_indent(self, ctx):
+        if ctx.pipe_separated:
+            return ' | ', '   | '
+        separator = ' ' * ctx.txt_separating_spaces
+        return separator, separator
+
+    def write(self, tokens, indent):
+        indent = indent * self.indent_string
+        values = [t.value for t in tokens]
+        row = indent + self.separator.join(values)
+        if self.pipes:
+            row = '| ' + row + ' |'
+        else:
+            row = row.rstrip()
+        self.output.write(row)
+
+    def write_newline(self):
+        self.output.write('\n')
